backend/requirements.txt
asgiref==3.11.1
Django==6.0.2
pillow==12.1.1
sqlparse==0.5.5
djangorestframework==3.16.1
